finish the design for the postal code
