+++ src/index.html
@@ -15,5 +15,5 @@
     <div id="index"></div>
 </body>
 
-<script src="js/script.js"></script>
+<script type="text/javascript" src="js/script.js"></script>
 </html>
